+++ CHANGELOG.md
@@ -5,6 +5,20 @@ Changelog
 [Unreleased]
 ------------
 
+### Added
+
+ - Add method execute to Transition
+ - Item tracks state changes which can be released by Item#releaseRecordedStateChanges
+
+### Changed
+
+ - Transition can be initialized without a target step
+ - Deprecate executeActions and executePostActions of Transition
+ - Call Transition#execute in AbstractTransitionHandler
+ - Use Item#releaseRecordedStateChanges in RepositoryBasedTransitionHandler to store all state changes
+
+[2.0.2] (2019-12-05) 
+
 ### Fixed
 
  - Fix initial value of item state
@@ -51,6 +65,7 @@ Changelog
 
 
 [Unreleased]:  https://github.com/netzmacht/workfow/compare/2.0.1...hotfix/2.0.2
+[2.0.2]:       https://github.com/netzmacht/workfow/compare/2.0.1...2.0.2
 [2.0.1]:       https://github.com/netzmacht/workfow/compare/2.0.0...2.0.1
 [2.0.0]:       https://github.com/netzmacht/workfow/compare/2.0.0-beta1...2.0.0
 [2.0.0-beta1]: https://github.com/netzmacht/workfow/compare/1.0.0-beta2...2.0.0-beta1